Allow LLD to use and version _end.
LLD seems to replace all uses of _end in libc with the end of libc's
initialized data segment, not the main program's. Giving libc some
space to merge _end with that of the main program allows _end references
from within libc to be resolved correctly.

The only use I know for this is sbrk and brk.

Other archs probably need similar changes.