Use file-prefix-map to lie about build paths
Unconditionally remap the cross sysroot to /, so debug information
points to (for example) /usr/include rather than
/usr/obj/usr/src/amd64.amd64/tmp/usr/include.

If MK_REPRODUCIBLE_BUILD is on, then use file-prefix-map to replace
SRCDIR with /usr/src, and OBJDIR with /usr/obj/usr/src. This causes
builds to appear to come from /usr/src and be build from /usr/obj no
matter the real paths.